GOTH-598 Update newsletters (#278)
- When a user goes to the page they should get randomly assigned to either variant 0 or 1.

a cookie, "_experiment_no-duplicates-in-river", is set with a value of either 0 or 1. This value should be the same upon returning to the page (the cookie lasts for 30 days and this timer is refreshed on new visits).

The experiment name ("no-duplicates-in-river") and variant ("0" or "1") are attached to all GA events as 'experimentName' and 'experimentVariant' respectively.

If the variant is 0, the home page remains unchanged.
If the variant is 1, stories in the latest news section at the bottom of the homepage are deduplicated from the homepage topper. In other words, anything that appears in the "Latest" section in the homepage topper shouldn't also appear in the "Latest" river.

Because the latest section in the homepage topper is also deduplicated from the curated featured articles, there's also of a special case. If either or both of the curated articles in the homepage topper are among the top 6 latest articles, they'll also be deduplicated from the river at the bottom of the page.